move fastify to cli
