+++ packages/cli/package.json
@@ -18,6 +18,7 @@
     "command-exists": "^1.2.8",
     "commander": "^4.0.1",
     "console.table": "^0.10.0",
+    "fastify": "^2.13.1",
     "fs-extra": "^8.1.0",
     "glob": "^7.1.6",
     "glob-promise": "^3.4.0",
